For #1817: Multi-select toolbar changes when minimized
