+++ app/src/main/java/org/mozilla/fenix/library/bookmarks/BookmarkUIView.kt
@@ -4,9 +4,17 @@
 
 package org.mozilla.fenix.library.bookmarks
 
+import android.graphics.PorterDuff
+import android.graphics.PorterDuffColorFilter
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.ImageButton
 import android.widget.LinearLayout
+import androidx.appcompat.app.AppCompatActivity
+import androidx.appcompat.view.menu.ActionMenuItemView
+import androidx.appcompat.widget.ActionMenuView
+import androidx.appcompat.widget.Toolbar
+import androidx.core.content.ContextCompat
 import io.reactivex.Observable
 import io.reactivex.Observer
 import io.reactivex.functions.Consumer
@@ -15,6 +23,8 @@ import mozilla.appservices.places.BookmarkRoot
 import mozilla.components.concept.storage.BookmarkNode
 import mozilla.components.support.base.feature.BackHandler
 import org.mozilla.fenix.R
+import org.mozilla.fenix.ext.asActivity
+import org.mozilla.fenix.ext.getColorIntFromAttr
 import org.mozilla.fenix.mvi.UIView
 
 class BookmarkUIView(
@@ -28,12 +38,14 @@ class BookmarkUIView(
     var mode: BookmarkState.Mode = BookmarkState.Mode.Normal
         private set
 
-    var canGoBack = false
+    private var canGoBack = false
 
     override val view: LinearLayout = LayoutInflater.from(container.context)
         .inflate(R.layout.component_bookmark, container, true) as LinearLayout
 
     private val bookmarkAdapter: BookmarkAdapter
+    private val context = container.context
+    private val activity = context?.asActivity()
 
     init {
         view.bookmark_list.apply {
@@ -48,6 +60,10 @@ class BookmarkUIView(
         if (it.mode != mode) {
             mode = it.mode
             actionEmitter.onNext(BookmarkAction.ModeChanged)
+            when (val modeCopy = mode) {
+                is BookmarkState.Mode.Normal -> setUIForNormalMode(it.tree)
+                is BookmarkState.Mode.Selecting -> setUIForSelectingMode(modeCopy)
+            }
         }
     }
 
@@ -59,4 +75,89 @@ class BookmarkUIView(
     }
 
     fun getSelected(): Set<BookmarkNode> = bookmarkAdapter.selected
+
+    private fun setToolbarColors(foreground: Int, background: Int) {
+        val toolbar = (activity as AppCompatActivity).findViewById<Toolbar>(R.id.navigationToolbar)
+        val colorFilter = PorterDuffColorFilter(
+            ContextCompat.getColor(context, foreground), PorterDuff.Mode.SRC_IN
+        )
+        toolbar.setBackgroundColor(ContextCompat.getColor(context, background))
+        toolbar.setTitleTextColor(ContextCompat.getColor(context, foreground))
+
+        themeToolbar(
+            toolbar, foreground,
+            background, colorFilter
+        )
+    }
+
+    private fun setUIForSelectingMode(
+        mode: BookmarkState.Mode.Selecting
+    ) {
+        activity?.title = context.getString(R.string.bookmarks_multi_select_title, mode.selectedItems.size)
+        setToolbarColors(
+            R.color.white_color,
+            R.attr.accentBright.getColorIntFromAttr(context!!)
+        )
+    }
+
+    private fun setUIForNormalMode(root: BookmarkNode?) {
+        activity?.title =
+            if (root?.title in setOf(
+                    "root",
+                    null
+                )
+            ) context.getString(R.string.library_bookmarks) else root!!.title
+
+        setToolbarColors(
+            R.attr.primaryText.getColorIntFromAttr(context!!),
+            R.attr.foundation.getColorIntFromAttr(context)
+        )
+    }
+
+    private fun themeToolbar(
+        toolbar: Toolbar,
+        textColor: Int,
+        backgroundColor: Int,
+        colorFilter: PorterDuffColorFilter? = null
+    ) {
+        toolbar.setTitleTextColor(ContextCompat.getColor(context!!, textColor))
+        toolbar.setBackgroundColor(ContextCompat.getColor(context, backgroundColor))
+
+        if (colorFilter == null) {
+            return
+        }
+
+        toolbar.overflowIcon?.colorFilter = colorFilter
+        (0 until toolbar.childCount).forEach {
+            when (val item = toolbar.getChildAt(it)) {
+                is ImageButton -> item.drawable.colorFilter = colorFilter
+                is ActionMenuView -> themeActionMenuView(item, colorFilter)
+            }
+        }
+    }
+
+    private fun themeActionMenuView(
+        item: ActionMenuView,
+        colorFilter: PorterDuffColorFilter
+    ) {
+        (0 until item.childCount).forEach {
+            val innerChild = item.getChildAt(it)
+            if (innerChild is ActionMenuItemView) {
+                themeChildren(innerChild, item, colorFilter)
+            }
+        }
+    }
+
+    private fun themeChildren(
+        innerChild: ActionMenuItemView,
+        item: ActionMenuView,
+        colorFilter: PorterDuffColorFilter
+    ) {
+        val drawables = innerChild.compoundDrawables
+        for (k in drawables.indices) {
+            drawables[k]?.let {
+                item.post { innerChild.compoundDrawables[k].colorFilter = colorFilter }
+            }
+        }
+    }
 }
